+++ scripts/start-server.js
@@ -1,4 +1,5 @@
 const storage=require('../server/runtime-storage');
 storage.assertReady();
+require('./test-reset-startup').applyPending(storage);
 storage.log();
 require('../server/server');